新增评论功能

+++ src/main/java/com/web/service/PaperService.java
@@ -75,7 +75,7 @@ public class PaperService {
         paperMapper.updateByPrimaryKey(paper);
     }
 
-    public void updatePrice(Integer paperId, BigDecimal price) {
+    public void updatePrice(Integer paperId, Integer price) {
         Paper paper = selectById(paperId);
         paper.setPrice(price);
         paperMapper.updateByPrimaryKey(paper);
